+++ app/src/main/java/org/simple/clinic/instantsearch/InstantSearchEvent.kt
@@ -2,6 +2,7 @@ package org.simple.clinic.instantsearch
 
 import org.simple.clinic.facility.Facility
 import org.simple.clinic.patient.PatientSearchResult
+import java.util.UUID
 
 sealed class InstantSearchEvent
 
@@ -12,3 +13,7 @@ data class AllPatientsLoaded(val patients: List<PatientSearchResult>) : InstantS
 data class SearchResultsLoaded(val patientsSearchResults: List<PatientSearchResult>) : InstantSearchEvent()
 
 data class SearchQueryValidated(val result: InstantSearchValidator.Result) : InstantSearchEvent()
+
+data class SearchResultClicked(val patientId: UUID) : InstantSearchEvent()
+
+data class SearchQueryChanged(val searchQuery: String) : InstantSearchEvent()