+++ CHANGELOG.md
@@ -10,8 +10,8 @@
 - Add convenience classes for creating screens for the new navigation framework
 - [In Progress: 21 Dec 2020] Instant search in Faster registration
 - [In Progress: 23 Dec 2020] Migrate app to use ViewBinding
-- [In Progress: 31 Dec 2020] Add Instant search screen
 - [In Progress: 04 Jan 2020] Migrate `AuthenticationActivity` to the new navigation framework
+- [In Progress: 05 Jan 2021] Add Instant search screen
 
 ### Fixes
 - Fix BPpassport prefill value to have display value